[chore] plus 페이지 수정
